red-social-web/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.1
django-allauth==64.0.0
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.7
mysqlclient==2.2.4
numpy==2.1.0
oauthlib==3.2.2
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
wheel==0.44.0
